api/requirements.txt
fastapi==0.115.0
supabase==2.7.0
python-jose[cryptography]==3.3.0
pydantic==2.9.0
python-dotenv==1.0.1
httpx==0.27.0
